sample entry factories
