+++ src/main/java/org/jcodec/containers/mp4/muxer/PCMMP4MuxerTrack.java
@@ -1,5 +1,4 @@
 package org.jcodec.containers.mp4.muxer;
-import org.jcodec.api.NotSupportedException;
 import org.jcodec.common.Assert;
 import org.jcodec.common.AudioFormat;
 import org.jcodec.common.LongArrayList;
@@ -32,7 +31,6 @@ import org.jcodec.containers.mp4.boxes.TrakBox;
 import java.io.IOException;
 import java.lang.IllegalStateException;
 import java.nio.ByteBuffer;
-import java.nio.ByteOrder;
 import java.util.Date;
 
 /**
@@ -58,27 +56,12 @@ public class PCMMP4MuxerTrack extends AbstractMP4MuxerTrack {
         this.out = out;
         this.frameDuration = 1;
         this.frameSize = (format.getSampleSizeInBits() >> 3) * format.getChannels();
-        addSampleEntry(_audioSampleEntry(format));
+        addSampleEntry(AudioSampleEntry.audioSampleEntryPCM(format));
         this._timescale = format.getSampleRate();
 
         setTgtChunkDuration(new Rational(1, 2), Unit.SEC);
     }
     
-    public static AudioSampleEntry _audioSampleEntry(AudioFormat format) {
-        return MP4Muxer.audioSampleEntry(lookupFourcc(format), 1, format.getSampleSizeInBits() >> 3,
-                format.getChannels(), (int) format.getSampleRate(),
-                format.isBigEndian() ? ByteOrder.BIG_ENDIAN : ByteOrder.LITTLE_ENDIAN);
-    }
-    
-    public static String lookupFourcc(AudioFormat format) {
-        if (format.getSampleSizeInBits() == 16 && !format.isBigEndian())
-            return "sowt";
-        else if (format.getSampleSizeInBits() == 24)
-            return "in24";
-        else
-            throw new NotSupportedException("Audio format " + format + " is not supported.");
-    }
-
     @Override
     public void addFrame(Packet outPacket) throws IOException {
         addSamples(outPacket.getData().duplicate());
